added doubles

+++ include/declarations.h
@@ -6,7 +6,7 @@
 /*   By: opelser <opelser@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/08/14 20:59:57 by opelser       #+#    #+#                 */
-/*   Updated: 2023/08/17 13:31:25 by opelser       ########   odam.nl         */
+/*   Updated: 2023/08/17 22:56:42 by opelser       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,13 +48,13 @@ typedef struct		s_map
  */
 typedef struct		s_player
 {
-	float			x;
-	float			y;
+	double			x;
+	double			y;
 
-	float			delta_x;
-	float			delta_y;
+	double			delta_x;
+	double			delta_y;
 
-	float			angle;
+	double			angle;
 }					t_player;
 
 /**
